backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
email-validator>=2.0.0
bcrypt>=4.0.0
requests>=2.31.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0 